refactor: AdminController와 UserController로 Contoroller 이름을 변경
