+++ src/main/java/kakao/festapick/festival/controller/FestivalUserController.java
@@ -1,11 +1,19 @@
 package kakao.festapick.festival.controller;
 
 import java.util.List;
+import kakao.festapick.festival.dto.CustomFestivalRequestDto;
+import kakao.festapick.festival.dto.FestivalRequestDto;
 import kakao.festapick.festival.dto.FestivalResponseDto;
 import kakao.festapick.festival.service.FestivalService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -21,27 +29,59 @@ public class FestivalUserController {
         this.festivalService = festivalService;
     }
 
+    //FESTIVAL_MANAGER만
+    @PostMapping
+    public ResponseEntity<Long> addFestival(
+            @AuthenticationPrincipal String identifier,
+            @RequestBody CustomFestivalRequestDto requestDto
+    ) {
+        Long festivalId = festivalService.addCustomizedFestival(requestDto, identifier);
+        return new ResponseEntity<>(festivalId, HttpStatus.CREATED);
+    }
+
+    //해당 지역에서 열리는 모든 축제
     @GetMapping("/areacode/{areaCode}")
     public ResponseEntity<List<FestivalResponseDto>> getFestivalByArea(@PathVariable String areaCode){
         List<FestivalResponseDto> festivalResponseDtos = festivalService.findApprovedOneByArea(areaCode);
         return ResponseEntity.ok(festivalResponseDtos);
     }
 
+    //해당 지역에서 현재 열리고 있는 축제
     @GetMapping("/areacode/{areaCode}/current")
     public ResponseEntity<List<FestivalResponseDto>> getCurrentFestivalByArea(@PathVariable String areaCode){
-        List<FestivalResponseDto> festivalResponseDtos = festivalService.findByAreaAndDate(areaCode);
+        List<FestivalResponseDto> festivalResponseDtos = festivalService.findApprovedAreaAndDate(areaCode);
         return ResponseEntity.ok(festivalResponseDtos);
     }
 
+    //모든 지역의 축제 조회(승인된 축제만)
     @GetMapping("/all")
     public ResponseEntity<List<FestivalResponseDto>> getApprovedFestivals(){
         List<FestivalResponseDto> festivalResponseDtos = festivalService.findApproved();
         return ResponseEntity.ok(festivalResponseDtos);
     }
 
+    //Keyword를 통한 축제 검색
     @GetMapping
     public ResponseEntity<List<FestivalResponseDto>> getFestivalByKeyword(@RequestParam String keyword){
         List<FestivalResponseDto> festivalResponseDtos = festivalService.findApprovedOneByKeyword(keyword);
         return ResponseEntity.ok(festivalResponseDtos);
     }
+
+    //자신이 올린 축제에 대해서만 삭제 가능
+    @PatchMapping("/{festivalId}")
+    public ResponseEntity<FestivalResponseDto> updateFestivalInfo(
+            @PathVariable Long festivalId,
+            @RequestBody FestivalRequestDto requestDto
+    ){
+        FestivalResponseDto responseDto =  festivalService.updateFestival(festivalId, requestDto);
+        return ResponseEntity.ok(responseDto);
+    }
+
+    //자신이 올린 축제에 대해서만 삭제 가능
+    @DeleteMapping("/{festivalId}")
+    public ResponseEntity<Void> removeFestival(@PathVariable Long festivalId){
+        festivalService.removeOne(festivalId);
+        return ResponseEntity.noContent().build();
+    }
+
 }
